change vertex_descriptor container

+++ Graph.h
@@ -27,7 +27,7 @@ class Graph {
         // typedefs
         // --------
 
-        typedef unsigned long vertex_descriptor;
+        typedef unsigned int vertex_descriptor; 
         typedef std::pair<vertex_descriptor, vertex_descriptor> edge_descriptor;
 
         typedef std::vector<vertex_descriptor>::const_iterator vertex_iterator;
